# apispecs/src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Web Project API
  description: Api connect to Database
  version: 0.0.1
  license:
    name: TDTU Web Project
servers:
  - url: http://XXXXX/v1
    description: Develop server
  - url: http://XXXXX.com
    description: Product server
paths:
  /v1/searchUser:
    post:
      tags:
        - User
      summary: Api return list user with condition in request body
      operationId: searchUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSearchRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: User list form database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersSearchResponse'
  /v1/updateUser:
    post:
      tags:
        - User
      summary: Api for update info of user
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status update user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUpdateResponse'
  /v1/createUser:
    post:
      tags:
        - User
      summary: Api for create user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status update user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
  /v1/getUser:
    get:
      tags:
        - User
      summary: Api for get user by UserId
      operationId: getUser
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            pattern: '^[0-9]{1,19}$'
            example: '123456'
          required: true
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGetResponse'
  /v1/deleteUser:
    delete:
      tags:
        - User
      summary: Api for delete user
      operationId: deleteUser
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            pattern: '^[0-9]{1,19}$'
            example: '123456'
          required: true
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status delete user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDeleteResponse'
  /v1/changePassword:
    post:
      tags:
        - User
      summary: Api for change password user
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status update user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePassResponse'
...
components:
...
  schemas:
    WebApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: array
          items:
            $ref: '#/components/schemas/WebApiErrorDetails'
    WebApiErrorDetails:
      type: object
      properties:
        field:
          type: string
        errorCd:
          type: string
        message:
          type: string
        errorIdList:
          type: array
          items:
            type: integer
            example: 1
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 400
        message:
          type: string
    UserSearchRequest:
      type: object
      properties:
        email:
          description: Email of user
          type: string
          example: 'username@domain.com'
          pattern: '^\w+([.\-]?\w+)*@\w+([.\-]?\w+)*(\.\w{2,3})+$'
          maxLength: 200
        stdId:
          description: Student ID of user
          type: string
          example: '123456'
          maxLength: 200
        userName:
          description: Name of user
          type: string
          example: 'user'
          maxLength: 200
        address:
          description: Address of user
          type: string
          example: '164, Tan My'
          maxLength: 200
        idCard:
          description: ID Card number of user
          type: string
          example: '123456'
          maxLength: 200
        phone:
          description: Phone number of user
          type: string
          example: '123456'
          pattern: '^(84|0[3|5|7|8|9])+([0-9]{8})\b$'
          maxLength: 10
        offset:
          description: The offset rows before beginning to return the rows
          type: integer
          format: int32
          minimum: 0
          example: 1
        limit:
          description: The number of rows
          type: integer
          format: int32
          minimum: 1
          example: 20
    UsersSearchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            users:
              description: Users list
              type: array
              items:
                $ref: '#/components/schemas/UserResponse'
            resultsTotalCount:
              description: Number of user with condition
              type: integer
              format: int64
              example: 1000
    UserResponse:
      type: object
      properties:
        userId:
          description: ID of user
          type: string
          example: '123456'
        userName:
          description: Name of user
          type: string
          example: 'user'
        approveStatus:
          description: Status of user
          type: string
          example: 'APPROVED'
        stdId:
          description: Student ID of user
          type: string
          example: '123456'
        address:
          description: Address of user
          type: string
          example: '164, Tan My'
        idCard:
          description: ID card of user
          type: string
          example: '123456'
        fCard:
          description: Front of the user's identity card photo
          type: string
          example: '123456'
        bCard:
          description: Back of the user's identity card photo
          type: string
          example: '123456'
        portrait:
          description: User portrait photo
          type: string
          example: '123456'
        phone:
          description: Phone of user
          type: string
          example: '0900000000'
        email:
          description: Email of user
          type: string
          example: '0900000000'
        password:
          description: Password of user
          type: string
          example: '************'
        roleId:
          description: roleId of user
          type: string
          example: '123456'
        roleName:
          description: roleName of user
          type: string
          example: '123456'
        createUserId:
          description: Id of user create
          type: string
          example: 'user'
        createUserName:
          description: Name of user create
          type: string
          example: 'user'
        createDatetime:
          description: Date time create user
          type: string
          example: '2020/04/01'
        lastupUserId:
          description: Id of user update
          type: string
          example: 'user'
        lastupUserName:
          description: Name of user update
          type: string
          example: 'user'
        lastupDatetime:
          description: Last date time update user
          type: string
          example: '2020/04/01'
    UserUpdateRequest:
      type: object
      required:
        - userId
        - userName
        - lastupUserId
      properties:
        userId:
          description: ID of user
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        lastupUserId:
          description: ID of user to create
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        stdId:
          description: Student ID of user
          type: string
          example: '123456'
          maxLength: 64
        userName:
          description: Name of user
          type: string
          example: 'user'
          maxLength: 64
        address:
          description: Address of user
          type: string
          example: '164, Tan My'
          maxLength: 200
        idCard:
          description: ID Card number of user
          type: string
          example: '123456'
          maxLength: 64
        fCard:
          description: Front of the user's identity card photo
          type: string
          example: '123456'
          maxLength: 200
        bCard:
          description: Back of the user's identity card photo
          type: string
          example: '123456'
          maxLength: 200
        portrait:
          description: User portrait photo
          type: string
          example: '123456'
          maxLength: 200
    UserUpdateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the update user
          type: string
          example: 'OK'
    UserCreateRequest:
      type: object
      required:
        - email
        - password
        - roleId
        - createUserId
        - userName
        - phone
      properties:
        email:
          description: Email of user
          type: string
          example: 'username@domain.com'
          pattern: '^\w+([.\-]?\w+)*@\w+([.\-]?\w+)*(\.\w{2,3})+$'
          maxLength: 200
        password:
          description: Password of user
          type: string
          example: '123456'
          minLength: 6
          maxLength: 64
        roleId:
          description: ID of role
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        createUserId:
          description: ID of user to create
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        stdId:
          description: Student ID of user
          type: string
          example: '123456'
          maxLength: 64
        userName:
          description: Name of user
          type: string
          example: 'user'
          minLength: 1
          maxLength: 64
        address:
          description: Address of user
          type: string
          example: '164, Tan My'
          maxLength: 200
        idCard:
          description: ID Card number of user
          type: string
          example: '123456'
          maxLength: 64
        fCard:
          description: Front of the user's identity card photo
          type: string
          example: '123456'
          maxLength: 200
        bCard:
          description: Back of the user's identity card photo
          type: string
          example: '123456'
          maxLength: 200
        portrait:
          description: User portrait photo
          type: string
          example: '123456'
          maxLength: 200
        phone:
          description: Phone number of user
          type: string
          example: '123456'
          pattern: '^(84|0[3|5|7|8|9])+([0-9]{8})\b$'
          maxLength: 10
    UserCreateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the create user
          type: string
          example: 'OK'
    UserGetResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
    UserDeleteResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the delete user
          type: string
          example: 'OK'
    ChangePassRequest:
      type: object
      required:
        - userId
        - passwordOld
        - password
        - passwordConfirm
        - lastupUserId
      properties:
        userId:
          description: ID of user
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        lastupUserId:
          description: ID of user to create
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        passwordOld:
          description: Password of user
          type: string
          example: '123456'
          minLength: 6
          maxLength: 64
        password:
          description: Password of user
          type: string
          example: '123456'
          minLength: 6
          maxLength: 64
        passwordConfirm:
          description: Password of user
          type: string
          example: '123456'
          minLength: 6
          maxLength: 64
    ChangePassResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the change pass
          type: string
          example: 'OK'
...
  responses:
    webApiValidationError:
      description: webApi input value validation failure
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/WebApiErrorResponse'
            properties:
              status:
                example: 400
...
    401:
      description: Restrict access
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 401
              error:
                example: Unauthorized
    403:
      description: Access request denied
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 403
              error:
                example: Forbidden
    404:
      description: Page not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 404
              error:
                example: Not Found
    500:
      description: The server is having problems
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 500
              error:
                example: Internal Server Error
...